Refactor JSON exporter
This commit refactors the JSON exporter to avoid typing the objects out
in full. Instead we rely on a new mapValues method which applies some
function to every Cereal in a given Cereals.

As well as removing a bunch of code, mapValues will be reused to build
the same objects in order to export to Nix.