Add Nix export format
This commit adds a Nix export format, with structure based on the JSON
export.

The toNixExpression function here is extremely limited, but it should be
sufficient to produce the limited subset of Nix that Root Loops actually
requires.

Although the majority of toNix is a straight copy of toJSON, I didn't
think it was worth factoring out any more of the commonalities than
mapValues already does.